requirements.txt
revChatGPT~=4.0.1
fastapi~=0.92.0
uvicorn[standard]~=0.22.0
PyMySQL~=1.0.2
DBUtils~=3.0.2
pydantic~=1.10.5
starlette~=0.25.0
python-dotenv~=1.0.0
sse-starlette~=0.4.0
aiomysql~=0.0.21
itsdangerous
